requirements.txt
altgraph==0.17.2
cycler==0.11.0
fonttools==4.29.1
future==0.18.2
iso8601==1.0.2
kiwisolver==1.3.2
matplotlib==3.3.3
numpy==1.22.2
packaging==21.3
pefile==2021.9.3
Pillow==9.0.1
pyinstaller==4.9
pyinstaller-hooks-contrib==2022.2
pymodbus==2.5.3
pyparsing==3.0.7
pyserial==3.5
python-dateutil==2.8.2
pywin32-ctypes==0.2.0
PyYAML==6.0
serial==0.0.97
six==1.16.0
